LayoutTests/performance-api/performance-observer-exception-expected.txt

CONSOLE MESSAGE: line 23: EXCEPTION MESSAGE IN CALLBACK
Ensure PerformanceObserver Exceptions are propogated.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PerformanceObserver callback fired
PASS onerror: EXCEPTION MESSAGE IN CALLBACK
PASS onerror: TypeError: Cannot call a class constructor without |new|
PASS successfullyParsed is true

TEST COMPLETE
